feat: refactor transaction modal setup callback communication minor naming corrections shift publisher link in the transaction modal correct variable names, show total fees in form of formatted sats signed-off-by: pavan joshi <pavanj914@gmail.com>
